+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## main
+### Changed
+- Renamed `PeerHandle::next_message()` to `recv_message()`.
 
 ## v0.4.2 - 2021-05-20
 ### Fixed